requirements.txt
flask
psycopg[binary]
bcrypt
python-dotenv